requirements.txt
agate==1.9.1
annotated-types==0.7.0
asttokens==3.0.1
attrs==25.4.0
babel==2.17.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
daff==1.4.2
dbt-adapters==1.22.5
dbt-common==1.37.2
dbt-core==1.11.2
dbt-extractor==0.6.0
dbt-postgres==1.10.0
dbt-protos==1.0.419
dbt-semantic-interfaces==0.9.0
debugpy==1.8.19
decorator==5.2.1
deepdiff==8.6.1
executing==2.2.1
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.1.0
greenlet==3.3.0
idna==3.11
importlib_metadata==8.7.1
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
leather==0.4.1
MarkupSafe==3.0.3
mashumaro==3.14
matplotlib==3.10.8
matplotlib-inline==0.2.1
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.1
opencv-python==4.13.0.90
orderly-set==5.5.0
packaging==25.0
pandas==2.3.3
parsedatetime==2.6
parso==0.8.5
pathspec==0.12.1
pillow==12.1.0
platformdirs==4.5.1
polars==1.37.1
polars-runtime-32==1.37.1
prompt_toolkit==3.0.52
protobuf==6.33.4
psutil==7.2.1
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
scipy==1.17.0
setuptools==80.9.0
six==1.17.0
snowplow-tracker==1.1.0
SQLAlchemy==2.0.45
sqlparse==0.5.4
stack-data==0.6.3
sympy==1.14.0
Telethon==1.42.0
text-unidecode==1.3
torch==2.9.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.5
ultralytics-thop==2.0.18
urllib3==2.6.3
wcwidth==0.2.14
wheel==0.45.1
zipp==3.23.0
